+++ CHANGELOG.md
@@ -50,6 +50,7 @@
   - `vagrant ssh -c` will now send stderr to stderr and stdout to stdout
     on the host machine, instead of all output to stdout.
   - Vagrant can now be interrupted during the "importing" step.
+  - NFS exports will no longer be cleared when an expected error occurs. [GH-577]
 
 ## 0.8.10 (December 10, 2011)